Clean up foreach code, create new SyntaxTokens Print class
closes #25
